Remove Attributes (contains) & ExactAttributes renamed to Attributes
